# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ONER API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Find Users",
        "description": "Accounts whose email contains `email`, ignoring case, newest first.",
        "operationId": "find_users_admin_users_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 320
                },
                {
                  "type": "null"
                }
              ],
              "title": "Email"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_UserAdmin_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{user_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Read User",
        "description": "An account with every course it owns and every payment it attempted.",
        "operationId": "read_user_admin_users__user_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAdminDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Edit User",
        "operationId": "edit_user_admin_users__user_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAdmin"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/purchases": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Find Purchases",
        "description": "Transaction history, newest first. Filter by part of the buyer's email,\nby course, or by status; pending rows are abandoned checkouts.",
        "operationId": "find_purchases_admin_purchases_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 320
                },
                {
                  "type": "null"
                }
              ],
              "title": "Email"
            }
          },
          {
            "name": "course_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Course Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PurchaseStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PurchaseAdmin_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/purchases/{purchase_id}/refund": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Refund",
        "description": "Record a refund made in FreedomPay's merchant cabinet. This sends no\nmoney: it marks the purchase refunded and, if asked, takes the course away.",
        "operationId": "refund_admin_purchases__purchase_id__refund_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "purchase_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Purchase Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseAdmin"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/lessons/{lesson_id}/video/upload": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Upload Video",
        "operationId": "upload_video_admin_lessons__lesson_id__video_upload_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoUploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoUploadTarget"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/lessons/{lesson_id}/video": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Lesson Video",
        "description": "Where the lesson's video stands. A pending upload is checked with\nKinescope and swapped in if processing has finished.",
        "operationId": "lesson_video_admin_lessons__lesson_id__video_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonVideo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/lessons/{lesson_id}/video/pending": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Discard Pending Video",
        "operationId": "discard_pending_video_admin_lessons__lesson_id__video_pending_delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Currency": {
        "type": "string",
        "enum": [
          "KGS",
          "KZT",
          "UZS",
          "USD"
        ],
        "title": "Currency"
      },
...
      "Gateway": {
        "type": "string",
        "enum": [
          "freedompay",
          "manual"
        ],
        "title": "Gateway"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LessonVideo": {
        "properties": {
          "video_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Id"
          },
          "pending_video_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Video Id"
          },
          "pending_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Status"
          },
          "state": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "in_progress",
                  "complete",
                  "failed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "duration": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          }
        },
        "type": "object",
        "title": "LessonVideo"
      },
...
      "OwnedCourseAdmin": {
        "properties": {
          "course_id": {
            "type": "integer",
            "title": "Course Id"
          },
          "course_title": {
            "type": "string",
            "title": "Course Title"
          },
          "granted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Granted At"
          },
          "source_purchase_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Purchase Id"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "course_title",
          "granted_at"
        ],
        "title": "OwnedCourseAdmin"
      },
      "Page_PurchaseAdmin_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PurchaseAdmin"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "Page[PurchaseAdmin]"
      },
      "Page_UserAdmin_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserAdmin"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "Page[UserAdmin]"
      },
      "PurchaseAdmin": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "user_email": {
            "type": "string",
            "title": "User Email"
          },
          "course_id": {
            "type": "integer",
            "title": "Course Id"
          },
          "course_title": {
            "type": "string",
            "title": "Course Title"
          },
          "gateway": {
            "$ref": "#/components/schemas/Gateway"
          },
          "gateway_txn_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gateway Txn Id"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "status": {
            "$ref": "#/components/schemas/PurchaseStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "refunded_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refunded At"
          },
          "refunded_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refunded By"
          },
          "refund_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refund Note"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "user_email",
          "course_id",
          "course_title",
          "gateway",
          "amount",
          "currency",
          "status",
          "created_at"
        ],
        "title": "PurchaseAdmin"
      },
      "PurchaseStatus": {
        "type": "string",
        "enum": [
          "pending",
          "paid",
          "failed",
          "refunded"
        ],
        "title": "PurchaseStatus"
      },
...
      "RefundRequest": {
        "properties": {
          "revoke_access": {
            "type": "boolean",
            "title": "Revoke Access"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revoke_access"
        ],
        "title": "RefundRequest",
        "description": "A refund already made in FreedomPay's merchant cabinet."
      },
...
      "UserAdmin": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "role",
          "is_active",
          "created_at"
        ],
        "title": "UserAdmin"
      },
      "UserAdminDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "owned_courses": {
            "items": {
              "$ref": "#/components/schemas/OwnedCourseAdmin"
            },
            "type": "array",
            "title": "Owned Courses",
            "default": []
          },
          "purchases": {
            "items": {
              "$ref": "#/components/schemas/PurchaseAdmin"
            },
            "type": "array",
            "title": "Purchases",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "role",
          "is_active",
          "created_at"
        ],
        "title": "UserAdminDetail"
      },
...
      "UserRole": {
        "type": "string",
        "enum": [
          "student",
          "admin"
        ],
        "title": "UserRole"
      },
      "UserUpdate": {
        "properties": {
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "password": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UserUpdate",
        "description": "What an admin can change on an account. The email stays its owner's, and\nany other field sent is a 422 rather than silently ignored."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VideoUploadRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename"
          },
          "filesize": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Filesize"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "filesize"
        ],
        "title": "VideoUploadRequest"
      },
      "VideoUploadTarget": {
        "properties": {
          "video_id": {
            "type": "string",
            "title": "Video Id"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          }
        },
        "type": "object",
        "required": [
          "video_id",
          "endpoint"
        ],
        "title": "VideoUploadTarget",
        "description": "Give `endpoint` to tus-js-client. It takes no token."
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      },
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}